Reorganize init command for better maintainability
