+++ .changeset/many-tools-melt.md
@@ -1,0 +1,5 @@
+---
+"@inlang/paraglide-js": patch
+---
+
+No longer add `identical-pattern` lint rule by default
