+++ inlang/source-code/paraglide/paraglide-js/src/cli/commands/init/index.ts
@@ -1,0 +1,1 @@
+export { initCommand as default } from "./command.js"
