+++ inlang/source-code/paraglide/paraglide-js/src/cli/main.ts
@@ -1,11 +1,11 @@
 import { Command } from "commander"
 import { compileCommand } from "./commands/compile.js"
-import { initCommand } from "./commands/init.js"
+import InitCommand from "./commands/init/index.js"
 import { version } from "./state.js"
 
 export const cli = new Command()
 	.name("paraglide-js")
 	.addCommand(compileCommand)
-	.addCommand(initCommand)
+	.addCommand(InitCommand)
 	.showHelpAfterError()
 	.version(version)
